songcrud/requirements.txt
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
tzdata==2022.5
djangorestframework==3.14.0